AMQ-125130 bug fix - Virtual destinations replication works incorrectly
