+++ activemq-broker/src/main/java/org/apache/activemq/replica/ReplicaPlugin.java
@@ -22,19 +22,12 @@ import org.apache.activemq.broker.BrokerPluginSupport;
 import org.apache.activemq.broker.BrokerService;
 import org.apache.activemq.broker.MutableBrokerFilter;
 import org.apache.activemq.broker.jmx.AnnotatedMBean;
-import org.apache.activemq.broker.ConnectionContext;
-import org.apache.activemq.broker.ProducerBrokerExchange;
 import org.apache.activemq.broker.region.CompositeDestinationInterceptor;
-import org.apache.activemq.broker.region.Destination;
-import org.apache.activemq.broker.region.DestinationFilter;
 import org.apache.activemq.broker.region.DestinationInterceptor;
 import org.apache.activemq.broker.region.RegionBroker;
 import org.apache.activemq.broker.region.policy.PolicyEntry;
 import org.apache.activemq.broker.region.policy.PolicyMap;
-import org.apache.activemq.broker.region.virtual.CompositeDestinationFilter;
-import org.apache.activemq.command.ActiveMQDestination;
 import org.apache.activemq.command.ActiveMQQueue;
-import org.apache.activemq.command.Message;
 import org.apache.activemq.replica.jmx.ReplicationJmxHelper;
 import org.apache.activemq.replica.jmx.ReplicationView;
 import org.apache.activemq.replica.storage.ReplicaFailOverStateStorage;
@@ -129,63 +122,6 @@ public class ReplicaPlugin extends BrokerPluginSupport {
         interceptors = Arrays.copyOf(interceptors, interceptors.length + 1);
         interceptors[interceptors.length - 1] = new ReplicaDestinationInterceptor((ReplicaSourceBroker)sourceBroker, roleManagementBroker);
         compositeInterceptor.setInterceptors(interceptors);
-    }
-
-    private static class ReplicaDestinationInterceptor implements DestinationInterceptor {
-
-        private final ReplicaSourceBroker sourceBroker;
-        private final ReplicaRoleManagementBroker roleManagementBroker;
-
-        public ReplicaDestinationInterceptor(ReplicaSourceBroker sourceBroker, ReplicaRoleManagementBroker roleManagementBroker) {
-            this.sourceBroker = sourceBroker;
-            this.roleManagementBroker = roleManagementBroker;
-        }
-
-        @Override
-        public Destination intercept(Destination destination) {
-            return new ReplicaDestinationFilter(destination, sourceBroker, roleManagementBroker);
-        }
-
-        @Override
-        public void remove(Destination destination) {
-        }
-
-        @Override
-        public void create(Broker broker, ConnectionContext context, ActiveMQDestination destination) throws Exception {
-        }
-    }
-
-    public static class ReplicaDestinationFilter extends DestinationFilter {
-
-        private final boolean nextIsComposite;
-        private final ReplicaSourceBroker sourceBroker;
-        private final ReplicaRoleManagementBroker roleManagementBroker;
-
-        public ReplicaDestinationFilter(Destination next, ReplicaSourceBroker sourceBroker, ReplicaRoleManagementBroker roleManagementBroker) {
-            super(next);
-            this.nextIsComposite = this.next != null && this.next instanceof CompositeDestinationFilter;
-            this.sourceBroker = sourceBroker;
-            this.roleManagementBroker = roleManagementBroker;
-        }
-
-        @Override
-        public void send(ProducerBrokerExchange producerExchange, Message messageSend) throws Exception {
-            if(ReplicaRole.source.equals(roleManagementBroker.getRole())) {
-                super.send(producerExchange, messageSend);
-                if(!nextIsComposite) {
-                    // don't replicate composite destination
-                    sourceBroker.replicateSend(producerExchange, messageSend);
-                }
-            } else {
-                if(nextIsComposite) {
-                    // we jump over CompositeDestinationFilter as we don't want to fan out composite destinations on the replica side
-                    ((CompositeDestinationFilter) getNext()).getNext().send(producerExchange, messageSend);
-                } else {
-                    super.send(producerExchange, messageSend);
-                }
-            }
-        }
-
     }
 
     private MutativeRoleBroker buildReplicaBroker(Broker broker, ReplicaFailOverStateStorage replicaFailOverStateStorage, WebConsoleAccessController webConsoleAccessController) {
